integrated react-native-rename, react-native-config
