+++ client/mobile/.env.dev
@@ -1,0 +1,1 @@
+API_URL=127.0.0.1:5000
