+++ client/mobile/.env.prod
@@ -1,0 +1,1 @@
+API_URL=https://goose-or-not.herokuapp.com
