+++ client/mobile/.env.staging
@@ -1,0 +1,1 @@
+API_URL=https://goose-or-not-stage.herokuapp.com
