+++ client/mobile/android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.client",
+    package = "com.gooseornot",
 )
 
 android_resource(
     name = "res",
-    package = "com.client",
+    package = "com.gooseornot",
     res = "src/main/res",
 )
