+++ client/mobile/android/app/src/main/java/com/gooseornot/MainActivity.java
@@ -1,4 +1,4 @@
-package com.client;
+package com.gooseornot;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
      */
     @Override
     protected String getMainComponentName() {
-        return "client";
+        return "GooseOrNot";
     }
 }
