+++ client/mobile/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "client",
-  "displayName": "client"
+  "name": "GooseOrNot",
+  "displayName": "Goose Or Not"
 }
