+++ client/mobile/ios/GooseOrNot-tvOS/Info.plist
