+++ client/mobile/ios/GooseOrNot-tvOSTests/Info.plist
