+++ client/mobile/ios/GooseOrNot.xcodeproj/xcshareddata/xcschemes/GooseOrNot.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "client.app"
-               BlueprintName = "client"
-               ReferencedContainer = "container:client.xcodeproj">
+               BuildableName = "GooseOrNot.app"
+               BlueprintName = "GooseOrNot"
+               ReferencedContainer = "container:GooseOrNot.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "clientTests.xctest"
-               BlueprintName = "clientTests"
-               ReferencedContainer = "container:client.xcodeproj">
+               BuildableName = "GooseOrNotTests.xctest"
+               BlueprintName = "GooseOrNotTests"
+               ReferencedContainer = "container:GooseOrNot.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -61,9 +61,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "clientTests.xctest"
-               BlueprintName = "clientTests"
-               ReferencedContainer = "container:client.xcodeproj">
+               BuildableName = "GooseOrNotTests.xctest"
+               BlueprintName = "GooseOrNotTests"
+               ReferencedContainer = "container:GooseOrNot.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -71,9 +71,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "client.app"
-            BlueprintName = "client"
-            ReferencedContainer = "container:client.xcodeproj">
+            BuildableName = "GooseOrNot.app"
+            BlueprintName = "GooseOrNot"
+            ReferencedContainer = "container:GooseOrNot.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -94,9 +94,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "client.app"
-            BlueprintName = "client"
-            ReferencedContainer = "container:client.xcodeproj">
+            BuildableName = "GooseOrNot.app"
+            BlueprintName = "GooseOrNot"
+            ReferencedContainer = "container:GooseOrNot.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -113,9 +113,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "client.app"
-            BlueprintName = "client"
-            ReferencedContainer = "container:client.xcodeproj">
+            BuildableName = "GooseOrNot.app"
+            BlueprintName = "GooseOrNot"
+            ReferencedContainer = "container:GooseOrNot.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
