+++ client/mobile/ios/GooseOrNot/AppDelegate.h
