+++ client/mobile/ios/GooseOrNot/Images.xcassets/AppIcon.appiconset/Contents.json
