+++ client/mobile/ios/GooseOrNot/Images.xcassets/Contents.json
