+++ client/mobile/ios/GooseOrNotTests/GooseOrNotTests.m
@@ -14,11 +14,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface clientTests : XCTestCase
+@interface GooseOrNotTests : XCTestCase
 
 @end
 
-@implementation clientTests
+@implementation GooseOrNotTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
