+++ client/mobile/ios/GooseOrNotTests/Info.plist
